+++ packages/kit/src/runtime/server/page/respond_with_error.js
@@ -80,7 +80,7 @@ export async function respond_with_error({ event, options, state, status, error,
 				csr: get_option([default_layout], 'csr') ?? true
 			},
 			status,
-			error: handle_error_and_jsonify(event, options, error),
+			error: await handle_error_and_jsonify(event, options, error),
 			branch,
 			fetched,
 			event,
@@ -96,7 +96,7 @@ export async function respond_with_error({ event, options, state, status, error,
 		return static_error_page(
 			options,
 			error instanceof HttpError ? error.status : 500,
-			handle_error_and_jsonify(event, options, error).message
+			(await handle_error_and_jsonify(event, options, error)).message
 		);
 	}
 }